refactor(26.3): Update accessors to Minecraft 26.3

+++ common/src/main/java/net/blay09/mods/excompressum/mixin/ConstantValueAccessor.java
@@ -1,11 +1,11 @@
 package net.blay09.mods.excompressum.mixin;
 
-import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
+import net.minecraft.world.level.storage.loot.providers.number.ints.ConstantValue;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(ConstantValue.class)
 public interface ConstantValueAccessor {
     @Accessor
-    float getValue();
+    int getValue();
 }
